requirements.txt
blinker                1.9.0
click                  8.2.1
colorama               0.4.6
Flask                  3.1.1
Flask-WTF              1.2.2
itsdangerous           2.2.0
Jinja2                 3.1.6
MarkupSafe             3.0.2
mysql-connector-python 9.3.0
pip                    25.1.1
Werkzeug               3.1.3
WTForms                3.2.1